requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
Deprecated==1.2.13
Flask==2.0.2
gcloud==0.18.3
googleapis-common-protos==1.54.0
httplib2==0.20.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
jwcrypto==1.0
MarkupSafe==2.0.1
oauth2client==4.1.3
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.12.0
pyparsing==3.0.7
Pyrebase4==4.5.0
python-jwt==3.3.2
requests==2.27.1
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
urllib3==1.26.8
Werkzeug==2.0.2
wrapt==1.13.3
